Renamed file to store hardcoded plant locations.
